test: isolate config file tests
